Remove unused "fraction" requirement
This package uses the `fractions` module, which is in the Python
standard library.  There seems to be no reason to pull in this library.

I considered keeping an empty `arbitrary-precision` extra around for
compatibility, but I don't think there's any point since neither of the
reverse-dependencies found on
https://www.wheelodex.org/projects/moarchiving/rdepends/ appear to use
that extra.

See https://bugs.debian.org/1075905.